requirements.txt
aws-cdk-lib==2.166.0
constructs>=10.0.0,<11.0.0
pytest==8.3.2
boto3==1.35.68
botocore==1.35.68